Fix compiler warnings on 32 bit platforms without "large file" support
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@666 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ mtools.h
@@ -83,7 +83,7 @@ typedef struct device {
 				 * partitions which may take up parts
 				 * of a track) */
 
-	uint32_t sector_size; /* Non-default sector size */
+	uint16_t sector_size; /* Non-default sector size */
 
 	const char *cfg_filename; /* used for debugging purposes */
 } device_t;
